-Slight modification to the 4 subclasses  private magicPoint from double to int -Added w2-uml diagram

+++ src/AncientMagicBook.java
@@ -1,8 +1,8 @@
 public class AncientMagicBook extends MagicBook
 {
-    private double magicPoint;
+    private int magicPoint;
 
-    public AncientMagicBook(String ID, String title, String author, double magicPoint)
+    public AncientMagicBook(String ID, String title, String author, int magicPoint)
     {
         super(ID, title, author);
         this.magicPoint = magicPoint;
